V2.0 !!!  check in changes for MPI types and MPI_STATUS_IGNORE.

+++ mpi-serial/list.c
@@ -377,7 +377,7 @@ void AP_list_print(char *str, pList list)
   cur=list->head;
   while(cur)
     {
-      printf("%d ",(int)cur->data);
+      printf("%d ",(long int)cur->data);
       cur=cur->next;
     }
 
@@ -404,7 +404,7 @@ void AP_list_revprint(char *str, pList list)
   cur=list->tail;
   while(cur)
     {
-      printf("%d ",(int)cur->data);
+      printf("%d ",(long int)cur->data);
       cur=cur->prev;
     }
